refactoring: ingredient and coffee recipe classes moved to the domain package
