+++ src/main/kotlin/lv/id/jc/machine/domain/Ingredient.kt
@@ -1,4 +1,4 @@
-package lv.id.jc.machine
+package lv.id.jc.machine.domain
 
 enum class Ingredient(val description: String, val unit: String) {
     Water("water", "ml"),
