+++ src/test/groovy/lv/id/jc/machine/domain/CoffeeSpec.groovy
@@ -1,12 +1,13 @@
-package lv.id.jc.machine
+package lv.id.jc.machine.domain
 
+import lv.id.jc.machine.domain.Coffee
 import spock.lang.Narrative
 import spock.lang.See
 import spock.lang.Specification
 import spock.lang.Subject
 import spock.lang.Title
 
-import static lv.id.jc.machine.Ingredient.*
+import static lv.id.jc.machine.domain.Ingredient.*
 
 @Title('Coffee drinks recipe and price')
 @Narrative('''
